# ipx/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Solana transaction API",
        "title": "SVMlab API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:33153",
    "basePath": "/",
    "paths": {
...
        "/svm/protocol/rent-exemption/mint": {
            "post": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "protocol"
                ],
                "summary": "Minimum balance for an SPL Token mint",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Chain name, e.g. solana",
                        "name": "X-Chain-Name",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Chain network, e.g. testnet",
                        "name": "X-Chain-Network",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/misc.RentExemptionMintResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/svm/protocol/rent-exemption/nonce": {
            "post": {
                "description": "A nonce account has to stay rent exempt to keep holding its nonce, so this is what one must be funded with. It is also the floor a partial withdrawal has to leave behind.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "protocol"
                ],
                "summary": "Minimum balance for a durable nonce account",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Chain name, e.g. solana",
                        "name": "X-Chain-Name",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Chain network, e.g. testnet",
                        "name": "X-Chain-Network",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/misc.RentExemptionNonceResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/svm/protocol/rent-exemption/public-key": {
            "post": {
                "description": "Reads the size from a live account, so the caller does not have to know the layout. Pass a byte count to the space endpoint instead when the account does not exist yet.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "protocol"
                ],
                "summary": "Minimum balance for an account of the same size as an existing one",
                "parameters": [
                    {
                        "description": "Account",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/misc.RentExemptionPublicKeyRequest"
                        }
                    },
                    {
                        "type": "string",
                        "description": "Chain name, e.g. solana",
                        "name": "X-Chain-Name",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Chain network, e.g. testnet",
                        "name": "X-Chain-Network",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/misc.RentExemptionPublicKeyResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/svm/protocol/rent-exemption/space": {
            "post": {
                "description": "Takes a byte count rather than an account kind, which covers layouts none of the named endpoints describe and sizes no live account holds yet.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "protocol"
                ],
                "summary": "Minimum balance for a size given directly",
                "parameters": [
                    {
                        "description": "Space",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/misc.RentExemptionSpaceRequest"
                        }
                    },
                    {
                        "type": "string",
                        "description": "Chain name, e.g. solana",
                        "name": "X-Chain-Name",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Chain network, e.g. testnet",
                        "name": "X-Chain-Network",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/misc.RentExemptionSpaceResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/svm/protocol/rent-exemption/stake": {
            "post": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "protocol"
                ],
                "summary": "Minimum balance for a stake account",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Chain name, e.g. solana",
                        "name": "X-Chain-Name",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Chain network, e.g. testnet",
                        "name": "X-Chain-Network",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/misc.RentExemptionStakeResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/svm/protocol/rent-exemption/system": {
            "post": {
                "description": "A wallet holds lamports and no data, so this is the floor any account must clear. It is also what a transfer to a previously unused address has to meet, since the transfer creates the account.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "protocol"
                ],
                "summary": "Minimum balance for a plain wallet",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Chain name, e.g. solana",
                        "name": "X-Chain-Name",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Chain network, e.g. testnet",
                        "name": "X-Chain-Network",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/misc.RentExemptionSystemResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/svm/protocol/rent-exemption/token": {
            "post": {
                "description": "Every token balance lives in an account of its own, so opening a position in a new token costs this much before any tokens move.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "protocol"
                ],
                "summary": "Minimum balance for an SPL Token account",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Chain name, e.g. solana",
                        "name": "X-Chain-Name",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Chain network, e.g. testnet",
                        "name": "X-Chain-Network",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/misc.RentExemptionTokenResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/svm/protocol/rent-exemption/vote": {
            "post": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "protocol"
                ],
                "summary": "Minimum balance for a validator vote account",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Chain name, e.g. solana",
                        "name": "X-Chain-Name",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Chain network, e.g. testnet",
                        "name": "X-Chain-Network",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/misc.RentExemptionVoteResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "misc.RentExemptionMintResponse": {
            "type": "object",
            "properties": {
                "lamports": {
                    "type": "string"
                },
                "sol": {
                    "type": "string"
                },
                "space": {
                    "type": "integer"
                }
            }
        },
        "misc.RentExemptionNonceResponse": {
            "type": "object",
            "properties": {
                "lamports": {
                    "type": "string"
                },
                "sol": {
                    "type": "string"
                },
                "space": {
                    "type": "integer"
                }
            }
        },
        "misc.RentExemptionPublicKeyRequest": {
            "type": "object",
            "properties": {
                "public_key": {
                    "type": "string",
                    "example": "EodYvwsT22JTdNmvCeC974WjPiVYcxvfGpYLJxnB3JqK"
                }
            }
        },
        "misc.RentExemptionPublicKeyResponse": {
            "type": "object",
            "properties": {
                "lamports": {
                    "type": "string"
                },
                "public_key": {
                    "type": "string"
                },
                "sol": {
                    "type": "string"
                },
                "space": {
                    "type": "integer"
                }
            }
        },
        "misc.RentExemptionSpaceRequest": {
            "type": "object",
            "properties": {
                "space": {
                    "type": "string",
                    "example": "165"
                }
            }
        },
        "misc.RentExemptionSpaceResponse": {
            "type": "object",
            "properties": {
                "lamports": {
                    "type": "string"
                },
                "sol": {
                    "type": "string"
                },
                "space": {
                    "type": "integer"
                }
            }
        },
        "misc.RentExemptionStakeResponse": {
            "type": "object",
            "properties": {
                "lamports": {
                    "type": "string"
                },
                "sol": {
                    "type": "string"
                },
                "space": {
                    "type": "integer"
                }
            }
        },
        "misc.RentExemptionSystemResponse": {
            "type": "object",
            "properties": {
                "lamports": {
                    "type": "string"
                },
                "sol": {
                    "type": "string"
                },
                "space": {
                    "type": "integer"
                }
            }
        },
        "misc.RentExemptionTokenResponse": {
            "type": "object",
            "properties": {
                "lamports": {
                    "type": "string"
                },
                "sol": {
                    "type": "string"
                },
                "space": {
                    "type": "integer"
                }
            }
        },
        "misc.RentExemptionVoteResponse": {
            "type": "object",
            "properties": {
                "lamports": {
                    "type": "string"
                },
                "sol": {
                    "type": "string"
                },
                "space": {
                    "type": "integer"
                }
            }
        },
...
    }
}